requirement.txt
pandas
numpy
jupyter
selenium
requests
bs4
pyspark
apache-airflow
azure-cli